Updated imports and message ids in AmountTokens
