+++ src/modules/core/components/Fields/AmountTokens/index.ts
@@ -1,1 +1,1 @@
-export { default } from './AmountTokens';
+export { default, calculateFee } from './AmountTokens';
