+++ src/modules/dashboard/components/Dialogs/CreatePaymentDialog/CreatePaymentDialogForm.tsx
@@ -17,6 +17,7 @@ import { ItemDataType } from '~core/OmniPicker';
 import { ActionDialogProps } from '~core/Dialog';
 import DialogSection from '~core/Dialog/DialogSection';
 import { Select, Annotations, AmountTokens } from '~core/Fields';
+import { calculateFee } from '~core/Fields/AmountTokens';
 import Heading from '~core/Heading';
 import SingleUserPicker, { filterUserSelection } from '~core/SingleUserPicker';
 import PermissionRequiredInfo from '~core/PermissionRequiredInfo';
@@ -116,29 +117,6 @@ const UserAvatar = HookedUserAvatar({ fetchUser: false });
 const supRenderAvatar = (address: Address, item: ItemDataType<AnyUser>) => (
   <UserAvatar address={address} user={item} size="xs" notSet={false} />
 );
-
-// NOTE: The equation to calculate totalToPay is as following (in Wei)
-// totalToPay = (receivedAmount + 1) * (feeInverse / (feeInverse -1))
-// The network adds 1 wei extra fee after the percentage calculation
-// For more info check out
-// https://github.com/JoinColony/colonyNetwork/blob/806e4d5750dc3a6b9fa80f6e007773b28327c90f/contracts/colony/ColonyFunding.sol#L656
-
-export const calculateFee = (
-  receivedAmount: string, // amount that the recipient finally receives
-  feeInverse: string,
-  decimals: number,
-): { feesInWei: string; totalToPay: string } => {
-  const amountInWei = moveDecimal(receivedAmount, decimals);
-  const totalToPayInWei = bigNumberify(amountInWei)
-    .add(1)
-    .mul(feeInverse)
-    .div(bigNumberify(feeInverse).sub(1));
-  const feesInWei = totalToPayInWei.sub(amountInWei);
-  return {
-    feesInWei: feesInWei.toString(),
-    totalToPay: moveDecimal(totalToPayInWei, -1 * decimals),
-  }; // NOTE: seems like moveDecimal does not have strict typing
-};
 
 const CreatePaymentDialogForm = ({
   back,
